requirements.txt
beautifulsoup4>=4.12.3
Jinja2>=3.1.4
python-dateutil>=2.9.0.post0
PyYAML>=6.0.2
requests>=2.32.3
pytest>=8.3.4
